+++ src/dotnet/APIView/APIViewWeb/Startup.cs
@@ -36,7 +36,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 using Microsoft.OpenApi.Models;
-using Newtonsoft.Json.Linq;
 using APIViewWeb.Services;
 
 namespace APIViewWeb
@@ -116,6 +115,7 @@ namespace APIViewWeb
 
             services.AddSingleton<IReviewManager, ReviewManager>();
             services.AddSingleton<IAPIRevisionsManager, APIRevisionsManager>();
+            services.AddSingleton<IAutoReviewService, AutoReviewService>();
             services.AddSingleton<ICommentsManager, CommentsManager>();
             services.AddSingleton<INotificationManager, NotificationManager>();
             services.AddSingleton<IEmailTemplateService, EmailTemplateService>();
@@ -247,11 +247,12 @@ namespace APIViewWeb
                                 response = await context.Backchannel.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, context.HttpContext.RequestAborted);
                                 response.EnsureSuccessStatusCode();
 
-                                var orgs = JArray.Parse(await response.Content.ReadAsStringAsync());
+                                var orgsJson = await response.Content.ReadAsStringAsync();
+                                using var orgsDoc = JsonDocument.Parse(orgsJson);
                                 var orgNames = new StringBuilder();
 
                                 bool isFirst = true;
-                                foreach (var org in orgs)
+                                foreach (var org in orgsDoc.RootElement.EnumerateArray())
                                 {
                                     if (isFirst)
                                     {
@@ -261,7 +262,10 @@ namespace APIViewWeb
                                     {
                                         orgNames.Append(",");
                                     }
-                                    orgNames.Append(org["login"]);
+                                    if (org.TryGetProperty("login", out var loginProperty))
+                                    {
+                                        orgNames.Append(loginProperty.GetString());
+                                    }
                                 }
 
                                 string msEmail = await GetMicrosoftEmailAsync(context);
@@ -303,11 +307,21 @@ namespace APIViewWeb
             services.AddSingleton<IAuthorizationHandler, SamplesRevisionOwnerRequirementHandler>();
             services.AddSingleton(x =>
             {
-                return new CosmosClient(Configuration["CosmosEndpoint"], new DefaultAzureCredential());
+                var cosmosClientOptions = new CosmosClientOptions
+                {
+                    UseSystemTextJsonSerializerWithOptions = new JsonSerializerOptions
+                    {
+                        PropertyNameCaseInsensitive = true,
+                        DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
+                        Converters = { new JsonStringEnumConverter() }
+                    }
+                };
+                
+                return new CosmosClient(Configuration["CosmosEndpoint"], CredentialProvider.GetAzureCredential(), cosmosClientOptions);
             });
             services.AddSingleton(x =>
             {
-                return new BlobServiceClient(new Uri(Configuration["StorageAccountUrl"]), new DefaultAzureCredential());
+                return new BlobServiceClient(new Uri(Configuration["StorageAccountUrl"]), CredentialProvider.GetAzureCredential());
             });
 
             services.AddHostedService<ReviewBackgroundHostedService>();
@@ -324,6 +338,8 @@ namespace APIViewWeb
             services.AddSignalR(options => {
                 options.EnableDetailedErrors = true;
                 options.MaximumReceiveMessageSize =  1024 * 1024;
+            }).AddJsonProtocol(options => {
+                options.PayloadSerializerOptions.PropertyNamingPolicy = JsonNamingPolicy.CamelCase;
             });
             services.AddSwaggerGen(options =>
             {
@@ -357,13 +373,16 @@ namespace APIViewWeb
             var respString = await response.Content.ReadAsStringAsync();
             try
             {
-                var emails = JArray.Parse(respString);
-                foreach (var email in emails)
+                using var emailsDoc = JsonDocument.Parse(respString);
+                foreach (var email in emailsDoc.RootElement.EnumerateArray())
                 {
-                    var address = email["email"]?.Value<string>();
-                    if (address != null && address.EndsWith("@microsoft.com", StringComparison.OrdinalIgnoreCase))
+                    if (email.TryGetProperty("email", out var emailProperty))
                     {
-                        return address;
+                        var address = emailProperty.GetString();
+                        if (address != null && address.EndsWith("@microsoft.com", StringComparison.OrdinalIgnoreCase))
+                        {
+                            return address;
+                        }
                     }
                 }
                 return null;