+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Tests/Tools/ReleasePlanToolTests.cs
@@ -13,11 +13,13 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
     {
         private TestLogger<ReleasePlanTool> logger;
         private IDevOpsService devOpsService;
+        private IGitHelper gitHelper;
         private IGitHubService gitHubService;
         private ITypeSpecHelper typeSpecHelper;
         private IUserHelper userHelper;
         private IEnvironmentHelper environmentHelper;
         private ReleasePlanTool releasePlanTool;
+        private IInputSanitizer inputSanitizer;
 
         [SetUp]
         public void Setup()
@@ -25,7 +27,8 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
 
             logger = new TestLogger<ReleasePlanTool>();
             devOpsService = new MockDevOpsService();
-            gitHubService = new Mock<IGitHubService>().Object;
+            gitHubService = new MockGitHubService();
+            inputSanitizer = new InputSanitizer();
 
             var typeSpecHelperMock = new Mock<ITypeSpecHelper>();
             typeSpecHelperMock.Setup(x => x.IsRepoPathForPublicSpecRepo(It.IsAny<string>())).Returns(true);
@@ -39,13 +42,19 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
             environmentHelperMock.Setup(x => x.GetBooleanVariable(It.IsAny<string>(), It.IsAny<bool>())).Returns(false);
             environmentHelper = environmentHelperMock.Object;
 
+            var gitHelperMock = new Mock<IGitHelper>();
+            gitHelperMock.Setup(x => x.GetBranchName(It.IsAny<string>())).Returns("testBranch");
+            gitHelper = gitHelperMock.Object;
+
             releasePlanTool = new ReleasePlanTool(
                 devOpsService,
+                gitHelper,
                 typeSpecHelper,
                 logger,
                 userHelper,
                 gitHubService,
-                environmentHelper);
+                environmentHelper,
+                inputSanitizer);
         }
 
         [Test]
@@ -134,11 +143,13 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
 
             var testReleasePlanTool = new ReleasePlanTool(
                 devOpsService,
+                gitHelper,
                 typeSpecHelper,
                 logger,
                 userHelper,
                 gitHubService,
-                environmentHelperMock.Object);
+                environmentHelperMock.Object,
+                inputSanitizer);
 
             var testCodeFilePath = "TypeSpecTestData/specification/testcontoso/Contoso.Management";
 
@@ -173,11 +184,13 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
 
             var testReleasePlanTool = new ReleasePlanTool(
                 devOpsService,
+                gitHelper,
                 typeSpecHelper,
                 logger,
                 userHelper,
                 gitHubService,
-                environmentHelperMock.Object);
+                environmentHelperMock.Object,
+                inputSanitizer);
 
             var testCodeFilePath = "TypeSpecTestData/specification/testcontoso/Contoso.Management";
 
@@ -303,6 +316,29 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
         {
             var updateStatus = await releasePlanTool.UpdateLanguageExclusionJustification(100, "This is a test justification for excluding certain languages.");
             Assert.That(updateStatus.Message, Does.Contain("Updated language exclusion justification in release plan"));
+        }
+
+        [Test]
+        public async Task Test_link_sdk_pull_request_to_release_plan()
+        {
+            var cases = new (string language, string pullRequestUrl)[]
+            {
+                ("Python", "https://github.com/Azure/azure-sdk-for-python/pull/12345"),
+                (".NET", "https://github.com/Azure/azure-sdk-for-net/pull/12345"),
+                ("dotnet", "https://github.com/Azure/azure-sdk-for-net/pull/12345"),
+                ("Dotnet", "https://github.com/Azure/azure-sdk-for-net/pull/12345"),
+                ("csharp", "https://github.com/Azure/azure-sdk-for-net/pull/12345"),
+                ("Javascript", "https://github.com/Azure/azure-sdk-for-js/pull/12345"),
+                ("typescript", "https://github.com/Azure/azure-sdk-for-js/pull/12345"),
+                ("Java", "https://github.com/Azure/azure-sdk-for-java/pull/12345"),
+                ("Go", "https://github.com/Azure/azure-sdk-for-go/pull/12345"),
+            };
+
+            foreach (var (language, pullRequestUrl) in cases)
+            {
+                var response = await releasePlanTool.LinkSdkPullRequestToReleasePlan(language, pullRequestUrl, 1, 1);
+                Assert.That(response.Details, Has.Some.Contains("Successfully linked pull request to release plan"), $"Assertion failed for language '{language}' and PR '{pullRequestUrl}'.");
+            }
         }
     }
 }