+++ tools/js-sdk-release-tools/src/test/common/utils.test.ts
@@ -214,7 +214,7 @@ describe("getReleaseStatus", () => {
             apiVersion: "",
             sdkReleaseType: "stable",
         });
-        expect(result).toBe(false);
+        expect(result).toBe(true);
     });
 });